Added documentation and removed some redundent code.

+++ CHADOWN/CHADOWN.cs
@@ -6,7 +6,7 @@ using System.Windows.Forms;
 
 namespace CHADOWN
 {
-    public static class CHADMethods //Class containing all the functional methods for CHADOWN
+    public static class CHADMethods //Class containing all the functional methods for CHADOWN.
     {
         public static void AVComboMethod(string OutputDir) //Main part of the A/V Combo module.
         {
@@ -40,7 +40,7 @@ namespace CHADOWN
             }
             Console.Clear();
             Console.WriteLine("What would you like the file to be named? Don't include the file extension in the name.");
-            string OutputName = Console.ReadLine(); //Video file output is named
+            string OutputName = Console.ReadLine(); //Video file output is named.
             string AVCInputs = $"/k ffmpeg -i \"{Videofileselect.FileName}\" -i \"{Audiofileselect.FileName}\" -c copy \"{OutputDir}\\{OutputName}.mp4\"";
             Process p = Process.Start("CMD.exe", AVCInputs);
             p.WaitForExit();
@@ -76,7 +76,7 @@ namespace CHADOWN
             Process p = Process.Start("CMD.exe", strCmdText);
             p.WaitForExit();
         }
-        public static void FMPGmethod(string OutputDir) //Main part of the ffmpeg module
+        public static void FMPGmethod(string OutputDir) //Main part of the ffmpeg module.
         {
             Console.WriteLine("Please paste the link for the playlist (usually a .m3u8) file");
             string FMPGlink = Console.ReadLine();
@@ -97,12 +97,12 @@ namespace CHADOWN
         }
     }
    
-    public class ConfigIO//Handles creating, writing to, and rewriting the config
+    public class ConfigIO//Handles creating, writing to, and rewriting the config.
     {
         string ConfigPath = Directory.GetCurrentDirectory() + @"\" + "config.txt";
-        public void CreateConfig()//Creates the config if one is not found
+        public void CreateConfig()//Creates the config if one is not found.
         {
-            if (!File.Exists(ConfigPath))
+            if (!File.Exists(ConfigPath)) //If the config file doesn't exist, this creates it and assigns the user's Videos folder as the default.
             {
                 using (FileStream ConCreate = File.Create(ConfigPath))
                 {
@@ -111,7 +111,7 @@ namespace CHADOWN
                 }
             }
         }
-        public string ReadConfig()//Reads the output directory from the config file and copies it into memory
+        public string ReadConfig()//Reads the output directory from the config file and copies it into memory.
         {
             string Outpath = System.IO.File.ReadAllText(ConfigPath);
             return Outpath;
@@ -121,19 +121,17 @@ namespace CHADOWN
             string Outpath = System.IO.File.ReadAllText(ConfigPath);
             Console.Clear();
             Console.WriteLine("Select the new output directory.");
-            FolderBrowserDialog NewOutputDir = new FolderBrowserDialog(); //This opens a file dialog so the user can specify a new output directory.
-            if (NewOutputDir.ShowDialog() == System.Windows.Forms.DialogResult.OK)
+            FolderBrowserDialog NewOutputDir = new FolderBrowserDialog(); //Creates the object for finding the folder.
+            if (NewOutputDir.ShowDialog() == System.Windows.Forms.DialogResult.OK) //Opens dialog and user selects new output folder.
             {
                 Outpath = NewOutputDir.SelectedPath;
             };
-            if (NewOutputDir.ShowDialog() == DialogResult.OK)
-            {}
             Console.Clear();
-            if (File.Exists(ConfigPath))
+            if (File.Exists(ConfigPath)) //Finds the existing config file and deletes it.
             {
                 File.Delete(ConfigPath);
             }
-            using (FileStream ConCreate = File.Create(ConfigPath))
+            using (FileStream ConCreate = File.Create(ConfigPath)) //Creates and writes to the new config file.
             {
                 byte[] info = new UTF8Encoding(true).GetBytes(Outpath);
                 ConCreate.Write(info, 0, info.Length);
@@ -165,7 +163,7 @@ namespace CHADOWN
                 Console.WriteLine($"Current output dir is {outpath}");
                 string ConInputType = Console.ReadLine(); //User selects which one they want.
                 Console.Clear();
-                switch (ConInputType) //Allows the user to select which module they would like to use
+                switch (ConInputType) //Allows the user to select which module they would like to use.
                 {
                     case "A": CHADMethods.YTDLmethod(outpath); break;
                     case "B": CHADMethods.FMPGmethod(outpath); break;
@@ -175,7 +173,7 @@ namespace CHADOWN
                 }
                 Console.Clear();
                 Console.WriteLine("Continue? (y/n)");
-                if (Console.ReadLine() == "n") //The user chooses if they need to do another video
+                if (Console.ReadLine() == "n") //The user chooses if they need to do another video.
                 {
                     return; //Kill script by returning from main.
                 }
